+++ CHANGES.md
@@ -4,6 +4,7 @@
 - Added new setting "openInFullScreenMode", which states if the calendar should be rendered in full screen mode.
 - The "Enable browser notifications" options in the "Configuration" dialog is now tabbed in slightly (shows which is belongs to).
 - Added new search settings "left" and "top", which states the left/top positions that should be used for the "Search" dialog.
+- Added "Edit" button support on the event tooltips (next to the "Close" button).  They are now stacked and and CSS controlled.
 - 
 
 ### Version 1.3.3: